Update initial migration and adapt context for PostgreSQL
The initial migration has been updated to change from SQLite to PostgreSQL for better scaling and performance. This includes using jsonb types for IpInfo and ClientInfo for more flexibility in data representation and querying. The ShortUrl entity now includes an 'Id' property for consistency and better indexing. In UrlShortenerContext and migrations, Npgsql.EntityFrameworkCore.PostgreSQL.Metadata is utilized to use PostgreSQL's identity column behavior, and UseNpgsql is now utilized instead of UseSqlite.