Fix typos and update software versions in README
In this commit, several typos were corrected to improve readability. "Uisng" was corrected to "using". In addition to this, the unnecessary word "Minima" was removed. The API version was updated from "FastEndpoints(.NET 7)" to "ASP.NET core minimal API (.NET 7)", and the software prerequisites were updated from ".NET 7 Preview 4" and "Visual Studio 2022 Preview" to ".NET" and "Visual Studio 2022" respectively for clarity.